requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1
django-heroku==0.3.1
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
idna==2.10
lxml==4.5.2
psutil==5.7.2
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
zope.event==4.4
zope.interface==5.1.0
